Ensure correct attributes are always used for embed preview
