+++ packages/block-library/src/embed/util.js
@@ -7,8 +7,9 @@ import { DEFAULT_EMBED_BLOCK, WORDPRESS_EMBED_BLOCK, ASPECT_RATIOS } from './con
 /**
  * External dependencies
  */
-import { includes } from 'lodash';
+import { includes, kebabCase, toLower } from 'lodash';
 import classnames from 'classnames/dedupe';
+import memoize from 'memize';
 
 /**
  * WordPress dependencies
@@ -178,3 +179,39 @@ export function fallback( url, onReplace ) {
 		createBlock( 'core/paragraph', { content: renderToString( link ) } )
 	);
 }
+
+/***
+ * Gets block attributes based on the preview and responsive state.
+ *
+ * @param {Object} preview The preview data.
+ * @param {Object} currentClassNames The block's current class names.
+ * @param {boolean} isResponsive Boolean indicating if the block supports responsive content.
+ * @param {boolean} allowResponsive Apply responsive classes to fixed size content.
+ * @return {Object} Attributes and values.
+ */
+export const getAttributesFromPreview = memoize( ( preview, title, currentClassNames, isResponsive, allowResponsive = true ) => {
+	if ( ! preview ) {
+		return {};
+	}
+
+	const attributes = {};
+	// Some plugins only return HTML with no type info, so default this to 'rich'.
+	let { type = 'rich' } = preview;
+	// If we got a provider name from the API, use it for the slug, otherwise we use the title,
+	// because not all embed code gives us a provider name.
+	const { html, provider_name: providerName } = preview;
+	const providerNameSlug = kebabCase( toLower( '' !== providerName ? providerName : title ) );
+
+	if ( isFromWordPress( html ) ) {
+		type = 'wp-embed';
+	}
+
+	if ( html || 'photo' === type ) {
+		attributes.type = type;
+		attributes.providerNameSlug = providerNameSlug;
+	}
+
+	attributes.className = getClassNames( html, currentClassNames, isResponsive && allowResponsive );
+
+	return attributes;
+} );
